+++ base.js
@@ -386,10 +386,12 @@ export default class Base {
             if (!isNone(conf.default))
                 value = typeof conf.default === 'function' ? conf.default() : conf.default;
             if (isNone(value)) {
-                if (conf.optional)
-                    return { conf, value };
-                if (!conf.optional)
+                if (conf.optional || conf.required === false) return { conf, value };
+                if (conf.required === true || conf.optional === false)
                     throw buildError(RequiredError, `Missing required property at '${path}'`, this.validateType, path, null, value, "REQUIRED_PROPERTY_MISSING");
+                if (Base.autorequire || Base.autorequire === undefined)
+                    throw buildError(RequiredError, `Missing required property at '${path}'`, this.validateType, path, null, value, "REQUIRED_PROPERTY_MISSING");
+                return { conf, value };
             }
         }
         const isOfType = () => {